add test for get_airflow_connection_url
